+++ packages/core/src/definitions/styles.ts
@@ -64,7 +64,7 @@ export const builtInStyles: Partial<
     defaultValue: '0 0 0 0',
   },
   cfBackgroundColor: {
-    displayName: 'Background',
+    displayName: 'Background color',
     type: 'Text',
     group: 'style',
     description: 'The background color of the section',
@@ -82,7 +82,7 @@ export const builtInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The height of the section',
-    defaultValue: 'fill',
+    defaultValue: 'fit-content',
   },
   cfMaxWidth: {
     displayName: 'Max width',
@@ -110,7 +110,7 @@ export const builtInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The border of the section',
-    defaultValue: '0px outside rgba(255, 255, 255, 0)',
+    defaultValue: '1px solid rgba(0, 0, 0, 0)',
   },
   cfGap: {
     displayName: 'Gap',
@@ -120,13 +120,13 @@ export const builtInStyles: Partial<
     defaultValue: '0px',
   },
   cfBackgroundImageUrl: {
-    displayName: 'Background Image',
+    displayName: 'Background image',
     type: 'Text',
     defaultValue: '',
     description: 'Background image for section or container',
   },
   cfBackgroundImageScaling: {
-    displayName: 'Image Scaling',
+    displayName: 'Image scaling',
     type: 'Text',
     group: 'style',
     description: 'Adjust background image to fit, fill or tile the container',
@@ -149,7 +149,7 @@ export const builtInStyles: Partial<
     },
   },
   cfBackgroundImageAlignment: {
-    displayName: 'Image Alignment',
+    displayName: 'Image alignment',
     type: 'Text',
     group: 'style',
     description: 'Align background image to the edges of the container',
@@ -315,7 +315,7 @@ export const containerBuiltInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The margin of the container',
-    defaultValue: '0 auto 0 auto',
+    defaultValue: '0 Auto 0 Auto',
   },
 };
 
@@ -376,7 +376,7 @@ export const singleColumnBuiltInStyles: Partial<
     defaultValue: '0 0 0 0',
   },
   cfBackgroundColor: {
-    displayName: 'Background',
+    displayName: 'Background color',
     type: 'Text',
     group: 'style',
     description: 'The background color of the column',
@@ -401,7 +401,7 @@ export const singleColumnBuiltInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The border of the column',
-    defaultValue: '0px outside rgba(255, 255, 255, 0)',
+    defaultValue: '1px solid rgba(0, 0, 0, 0)',
   },
   cfGap: {
     displayName: 'Gap',
@@ -411,13 +411,13 @@ export const singleColumnBuiltInStyles: Partial<
     defaultValue: '0px',
   },
   cfBackgroundImageUrl: {
-    displayName: 'Background Image',
+    displayName: 'Background image',
     type: 'Text',
     defaultValue: '',
     description: 'Background image for section or container',
   },
   cfBackgroundImageScaling: {
-    displayName: 'Image Scaling',
+    displayName: 'Image scaling',
     type: 'Text',
     group: 'style',
     description: 'Adjust background image to fit, fill or tile the column',
@@ -440,7 +440,7 @@ export const singleColumnBuiltInStyles: Partial<
     },
   },
   cfBackgroundImageAlignment: {
-    displayName: 'Image Alignment',
+    displayName: 'Image alignment',
     type: 'Text',
     group: 'style',
     description: 'Align background image to the edges of the column',
@@ -452,8 +452,8 @@ export const singleColumnBuiltInStyles: Partial<
     group: 'style',
   },
   cfColumnSpanLock: {
-    type: 'Text',
-    defaultValue: 'false',
+    type: 'Boolean',
+    defaultValue: false,
     group: 'style',
   },
 };
@@ -466,7 +466,7 @@ export const columnsBuiltInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The margin of the columns',
-    defaultValue: '0 auto 0 auto',
+    defaultValue: '0 Auto 0 Auto',
   },
   cfWidth: {
     displayName: 'Width',
@@ -490,7 +490,7 @@ export const columnsBuiltInStyles: Partial<
     defaultValue: '10px 10px 10px 10px',
   },
   cfBackgroundColor: {
-    displayName: 'Background',
+    displayName: 'Background color',
     type: 'Text',
     group: 'style',
     description: 'The background color of the columns',
@@ -501,10 +501,10 @@ export const columnsBuiltInStyles: Partial<
     type: 'Text',
     group: 'style',
     description: 'The border of the columns',
-    defaultValue: '0px outside rgba(255, 255, 255, 0)',
+    defaultValue: '1px solid rgba(0, 0, 0, 0)',
   },
   cfBackgroundImageUrl: {
-    displayName: 'Background Image',
+    displayName: 'Background image',
     type: 'Text',
     defaultValue: '',
     description: 'Background image for section or container',
@@ -517,7 +517,7 @@ export const columnsBuiltInStyles: Partial<
     defaultValue: '10px 10px',
   },
   cfBackgroundImageScaling: {
-    displayName: 'Image Scaling',
+    displayName: 'Image scaling',
     type: 'Text',
     group: 'style',
     description: 'Adjust background image to fit, fill or tile the columns',
@@ -540,7 +540,7 @@ export const columnsBuiltInStyles: Partial<
     },
   },
   cfBackgroundImageAlignment: {
-    displayName: 'Image Alignment',
+    displayName: 'Image alignment',
     type: 'Text',
     group: 'style',
     description: 'Align background image to the edges of the columns',
@@ -552,8 +552,8 @@ export const columnsBuiltInStyles: Partial<
     group: 'style',
   },
   cfWrapColumns: {
-    type: 'Text',
-    defaultValue: 'false',
+    type: 'Boolean',
+    defaultValue: false,
     group: 'style',
   },
   cfWrapColumnsCount: {