+++ packages/core/src/fetchers/createExperience.ts
@@ -12,23 +12,33 @@ type createExperienceArgs = {
   mode?: ExternalSDKMode;
 };
 
-export const createExperience = ({
-  experienceEntry,
-  referencedAssets,
-  referencedEntries,
-  locale,
-}: createExperienceArgs): Experience<EntityStore> => {
-  if (!isExperienceEntry(experienceEntry)) {
-    throw new Error('Provided entry is not experience entry');
+/**
+ * Create an experience instance
+ * @param {string} json - JSON representation of the experience
+ */
+export function createExperience(json: string): Experience<EntityStore>;
+export function createExperience(args: createExperienceArgs): Experience<EntityStore>;
+
+export function createExperience(options: string | createExperienceArgs): Experience<EntityStore> {
+  if (typeof options === 'string') {
+    const entityStore = new EntityStore(options);
+    return {
+      entityStore,
+    };
+  } else {
+    const { experienceEntry, referencedAssets, referencedEntries, locale } = options;
+    if (!isExperienceEntry(experienceEntry)) {
+      throw new Error('Provided entry is not experience entry');
+    }
+
+    const entityStore = new EntityStore({
+      experienceEntry,
+      entities: [...referencedEntries, ...referencedAssets],
+      locale,
+    });
+
+    return {
+      entityStore,
+    };
   }
-
-  const entityStore = new EntityStore({
-    experienceEntry,
-    entities: [...referencedEntries, ...referencedAssets],
-    locale,
-  });
-
-  return {
-    entityStore,
-  };
-};
+}